imx-gpu-apitrace: Adjust work around for egltrace.so installation issue [MGS-2964]
The issue with egltrace.so seems intermittent. Make the workaround
conditional.

+++ imx/meta-bsp/recipes-graphics/imx-gpu-apitrace/imx-gpu-apitrace_7.1.0.bb
@@ -49,9 +49,11 @@ EXTRA_OECMAKE += \
                                                         '-DDISABLE_X11=ON', d), d)}"
 
 do_install_append() {
-    install -d ${D}/usr/lib/apitrace/wrappers
-    mv ${D}/usr/lib/arm-linux-gnueabi/apitrace/wrappers/egltrace.so ${D}/usr/lib/apitrace/wrappers/egltrace.so
-    rm -rf ${D}/usr/lib/arm-linux-gnueabi
+    if [ -f ${D}/usr/lib/arm-linux-gnueabi/apitrace/wrappers/egltrace.so ]; then
+        install -d ${D}/usr/lib/apitrace/wrappers
+        mv ${D}/usr/lib/arm-linux-gnueabi/apitrace/wrappers/egltrace.so ${D}/usr/lib/apitrace/wrappers/egltrace.so
+        rm -rf ${D}/usr/lib/arm-linux-gnueabi
+    fi
 }
 
 FILES_${PN} = "${bindir} ${libdir}"
